requirements.txt
google-cloud-aiplatform
google-adk
python-dotenv
python-docx
PyMuPDF
streamlit
google.adk.types